Fixes from code review on datamod with Kai
Add developer name to TODO

Use clang-format off/on instead off trailing slashes

Add a few more comments (more documentation still missing)

Remove oparator ValueType() on Entry - document modify (because it is
not obvious)

Remove Entry.has to Entry.isSet, remove Entry.isUnset

Remove BaseEntryParser and BaseEntryDumper

Remove EncodeMtg2Conf as MultiomOptions are not needed anymore

Rename MarsRecord to FullMarsRecord

Format share/generate_knowledge/*.py with `ruff format`

Add statType and more tests

Start documenting with doxygen

+++ src/multio/datamod/types/TypeOfStatisticalProcessing.h
@@ -1,5 +1,5 @@
 /*
- * (C) Copyright 1996- ECMWF.
+ * (C) Copyright 2025- ECMWF.
  *
  * This software is licensed under the terms of the Apache Licence Version 2.0
  * which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
@@ -19,8 +19,9 @@ namespace multio::datamod {
 
 //-----------------------------------------------------------------------------
 
-// TypeOfStatisticalProcessing defined with official integer representation in GRIB code table
-enum class TypeOfStatisticalProcessing : std::size_t
+// TypeOfStatisticalProcessing defined with official integer representation in GRIB 4.10 code table
+// https://codes.ecmwf.int/grib/format/grib2/ctables/4/10/
+enum class TypeOfStatisticalProcessing : std::int64_t
 {
     Average = 0,
     Accumulation = 1,
